fix(onboard): move preflight transition helper into boundary
- Adds `recordInitialPreflightTransition` on OnboardRuntimeBoundary
  so onboard.ts stays net-neutral for the codebase-growth guardrail.
- Updates `ONBOARD_MACHINE_EVENT_TYPES` fixture in transitions.test.ts
  to include the newly added `state.result.invalidated` event, fixing
  the transitions.test.ts:59 assertion failure in cli-test-shards (5).
- Drops the now-unused `advanceTo` import from onboard.ts.
