+++ src/lib/onboard/runtime-boundary.ts
@@ -8,6 +8,7 @@ import {
   shouldUpdateMachine,
 } from "../state/onboard-step-mutation";
 import type { OnboardStateFailedResult, OnboardStateResult } from "./machine/result";
+import { advanceTo } from "./machine/result";
 import { OnboardRuntime } from "./machine/runtime";
 import { assertValidOnboardMachineTransition } from "./machine/transitions";
 import type { OnboardMachineEventType, OnboardMachineState } from "./machine/types";
@@ -258,6 +259,37 @@ export class OnboardRuntimeBoundary {
     } = {},
   ): Promise<Session> {
     return this.getRuntime().emitRepairEvent(type, options);
+  }
+
+  /**
+   * Record the initial `init -> preflight` transition, honoring resume semantics.
+   * Fresh onboarding applies the transition; resumes invalidate stale replay
+   * results when the session has already advanced past `init`.
+   */
+  async recordInitialPreflightTransition(resume: boolean): Promise<void> {
+    const result = advanceTo("preflight", { metadata: { state: "init" } });
+    if (!resume) {
+      await this.recordStateResultWithStepCompatibility(result);
+      return;
+    }
+    const current = await this.getRuntime().session();
+    if (current.machine.state === result.next) {
+      await this.recordInvalidatedStateResult(result, {
+        reason: "already_at_target",
+        currentState: current.machine.state,
+        sourceState: "init",
+      });
+      return;
+    }
+    if (current.machine.state !== "init") {
+      await this.recordInvalidatedStateResult(result, {
+        reason: "source_state_mismatch",
+        currentState: current.machine.state,
+        sourceState: "init",
+      });
+      return;
+    }
+    await this.recordStateResultWithStepCompatibility(result);
   }
 
   async recordPostVerifyStarted(): Promise<Session> {
